requirements.txt
Django==1.10.1
djangorestframework==3.4.6
MySQL-python==1.2.5
Pillow==3.3.1
reportlab==3.3.0
requests==2.11.1
